+++ dart/math/lcp/newton/fischer_burmeister_newton_solver.cpp
@@ -408,7 +408,7 @@ LcpResult FischerBurmeisterNewtonSolver::solve(
 
   Eigen::VectorXd fastW;
   if (!options.warmStart
-      && detail::trySolveStrictInteriorStandardLcp(
+      && detail::trySolveStrictInteriorStandardLcpLltFirst(
           problem, absTol, std::max(absTol, compTol), x, &fastW)) {
     Eigen::VectorXd loEff;
     Eigen::VectorXd hiEff;