+++ dart/math/lcp/newton/minimum_map_newton_solver.cpp
@@ -371,7 +371,7 @@ LcpResult MinimumMapNewtonSolver::solve(
 
   Eigen::VectorXd fastW;
   if (!options.warmStart
-      && detail::trySolveStrictInteriorStandardLcp(
+      && detail::trySolveStrictInteriorStandardLcpLltFirst(
           problem, absTol, std::max(absTol, compTol), x, &fastW)) {
     Eigen::VectorXd loEff;
     Eigen::VectorXd hiEff;