+++ dart/math/lcp/newton/penalized_fischer_burmeister_newton_solver.cpp
@@ -447,7 +447,7 @@ LcpResult PenalizedFischerBurmeisterNewtonSolver::solve(
 
   Eigen::VectorXd fastW;
   if (!options.warmStart
-      && detail::trySolveStrictInteriorStandardLcp(
+      && detail::trySolveStrictInteriorStandardLcpLltFirst(
           problem, absTol, std::max(absTol, compTol), x, &fastW)) {
     Eigen::VectorXd loEff;
     Eigen::VectorXd hiEff;